moved ResizeProfilingBenchmarks
