+++ tests/ImageSharp.Tests/Processors/Filters/ResizeTests.cs
@@ -8,32 +8,6 @@ namespace ImageSharp.Tests
     using System.IO;
     using Processing;
     using Xunit;
-    using Xunit.Abstractions;
-
-    public class ResizeProfilingBenchmarks : MeasureFixture
-    {
-        public ResizeProfilingBenchmarks(ITestOutputHelper output)
-            : base(output)
-        {
-        }
-
-        public int ExecutionCount { get; set; } = 50;
-
-        [Theory]
-        [InlineData(100, 100)]
-        [InlineData(2000, 2000)]
-        public void ResizeBicubic(int width, int height)
-        {
-            this.Measure(this.ExecutionCount,
-                () =>
-                    {
-                        using (Image image = new Image(width, height))
-                        {
-                            image.Resize(width / 4, height / 4);
-                        }
-                    });
-        }
-    }
 
     public class ResizeTests : FileTestBase
     {
